Fixed issue 166, where two different coeditor sessions sometimes did not see each other's changes.
Superficially, rangy.saveSelection() was undefined and caused updates to not be processed by the remote event callback. The more underlying issue is that rangy.init() was never called - this is supposed to happen automatically when the page finishes loading, but sometimes it would not. The one line fix was to add a manual rangy.init() call when the text editor loads.