Improved CoTree maintaining consistency
The backing store is considered as a list of node which have a name and
a list of children. Adding/deleting nodes involves inserting a node ID
into the parent's list of children. Moving nodes involves deleting a
node ID from the previous parent and inserting the node ID in the new
parent's children list.